tests added to prove the thread safety of the money transfer
 * These tests have been added to proove that the money transfer
operation is thread safe.
 * There are two tests, one calls the thread safe money transfer method
and the other one calls the money transfer method which is not thread
safe.
 * Both the tests start with two accounts account-1 having 10000
balance and account-2 having 0 balance. Then 1 unit transfer is done
from account-1 to account-2 in 10000 threads.
 * After the execution of threads the thread-safe method has account-1
has 0 balance and the account-2 has 10000 balance.
 * Whereas in non thread-safe method testing the remaining balances are
always unpredictable and wrong because of the race-condition.
 *
 * I have used the method which is the thread safe one and this proves
the point.